book에서 복합검색 하기

+++ spring15book/src/main/java/com/kh/spring15/dto/BookDto.java
@@ -11,7 +11,7 @@ public class BookDto {
 	private String bookTitle;
 	private String bookAuthor;
 	private String bookPublicationDate;
-	private int bookPrice;
+	private float bookPrice;
 	private String bookPublisher;
 	private int bookPageCount;
 	private String bookGenre;
